Issue 80: TextContent extension for GFM tables, processed review comments
