--- commonmark-ext-gfm-tables-text/src/main/java/org/commonmark/ext/gfm/tables/text/TextTablesExtension.java
@@ -1,43 +1,0 @@
-package org.commonmark.ext.gfm.tables.text;
-
-import org.commonmark.Extension;
-import org.commonmark.ext.gfm.tables.internal.TableBlockParser;
-import org.commonmark.ext.gfm.tables.text.internal.TextTableNodeRenderer;
-import org.commonmark.parser.Parser;
-import org.commonmark.renderer.NodeRenderer;
-import org.commonmark.renderer.text.TextContentNodeRendererContext;
-import org.commonmark.renderer.text.TextContentNodeRendererFactory;
-import org.commonmark.renderer.text.TextContentRenderer;
-
-/**
- * Extension for GFM tables using "|" pipes (GitHub Flavored Markdown).
- * <p>
- * Create it with {@link #create()} and then configure it on the builders
- * ({@link org.commonmark.parser.Parser.Builder#extensions(Iterable)},
- * {@link TextContentRenderer.Builder#extensions(Iterable)}).
- * </p>
- */
-public class TextTablesExtension implements Parser.ParserExtension, TextContentRenderer.TextContentRendererExtension {
-
-    private TextTablesExtension() {
-    }
-
-    public static Extension create() {
-        return new TextTablesExtension();
-    }
-
-    @Override
-    public void extend(Parser.Builder parserBuilder) {
-        parserBuilder.customBlockParserFactory(new TableBlockParser.Factory());
-    }
-
-    @Override
-    public void extend(TextContentRenderer.Builder rendererBuilder) {
-        rendererBuilder.nodeRendererFactory(new TextContentNodeRendererFactory() {
-            @Override
-            public NodeRenderer create(TextContentNodeRendererContext context) {
-                return new TextTableNodeRenderer(context);
-            }
-        });
-    }
-}
