--- commonmark-ext-gfm-tables-text/src/main/javadoc/overview.html
@@ -1,6 +1,0 @@
-<html>
-<body>
-<b>Extension for rendering GFM tables as text using "|" pipes (GitHub Flavored Markdown)</b>
-<p>See {@link org.commonmark.ext.gfm.tables.text.TextTablesExtension}</p>
-</body>
-</html>
