+++ commonmark-ext-gfm-tables/src/main/java/org/commonmark/ext/gfm/tables/TablesExtension.java
@@ -2,12 +2,16 @@ package org.commonmark.ext.gfm.tables;
 
 import org.commonmark.Extension;
 import org.commonmark.ext.gfm.tables.internal.TableBlockParser;
-import org.commonmark.ext.gfm.tables.internal.TableNodeRenderer;
+import org.commonmark.ext.gfm.tables.internal.TableHtmlNodeRenderer;
+import org.commonmark.ext.gfm.tables.internal.TableTextContentNodeRenderer;
 import org.commonmark.renderer.html.HtmlRenderer;
 import org.commonmark.renderer.html.HtmlNodeRendererContext;
 import org.commonmark.renderer.html.HtmlNodeRendererFactory;
 import org.commonmark.parser.Parser;
 import org.commonmark.renderer.NodeRenderer;
+import org.commonmark.renderer.text.TextContentNodeRendererContext;
+import org.commonmark.renderer.text.TextContentNodeRendererFactory;
+import org.commonmark.renderer.text.TextContentRenderer;
 
 /**
  * Extension for GFM tables using "|" pipes (GitHub Flavored Markdown).
@@ -20,7 +24,8 @@ import org.commonmark.renderer.NodeRenderer;
  * The parsed tables are turned into {@link TableBlock} blocks.
  * </p>
  */
-public class TablesExtension implements Parser.ParserExtension, HtmlRenderer.HtmlRendererExtension {
+public class TablesExtension implements Parser.ParserExtension, HtmlRenderer.HtmlRendererExtension,
+        TextContentRenderer.TextContentRendererExtension {
 
     private TablesExtension() {
     }
@@ -39,7 +44,17 @@ public class TablesExtension implements Parser.ParserExtension, HtmlRenderer.Htm
         rendererBuilder.nodeRendererFactory(new HtmlNodeRendererFactory() {
             @Override
             public NodeRenderer create(HtmlNodeRendererContext context) {
-                return new TableNodeRenderer(context);
+                return new TableHtmlNodeRenderer(context);
+            }
+        });
+    }
+
+    @Override
+    public void extend(TextContentRenderer.Builder rendererBuilder) {
+        rendererBuilder.nodeRendererFactory(new TextContentNodeRendererFactory() {
+            @Override
+            public NodeRenderer create(TextContentNodeRendererContext context) {
+                return new TableTextContentNodeRenderer(context);
             }
         });
     }
