+++ pom.xml
@@ -24,7 +24,6 @@
         <module>commonmark-ext-autolink</module>
         <module>commonmark-ext-gfm-strikethrough</module>
         <module>commonmark-ext-gfm-tables</module>
-        <module>commonmark-ext-gfm-tables-text</module>
         <module>commonmark-ext-heading-anchor</module>
         <module>commonmark-ext-ins</module>
         <module>commonmark-ext-yaml-front-matter</module>
@@ -98,11 +97,6 @@
             <dependency>
                 <groupId>com.atlassian.commonmark</groupId>
                 <artifactId>commonmark-ext-gfm-tables</artifactId>
-                <version>0.10.1-SNAPSHOT</version>
-            </dependency>
-            <dependency>
-                <groupId>com.atlassian.commonmark</groupId>
-                <artifactId>commonmark-ext-gfm-tables-text</artifactId>
                 <version>0.10.1-SNAPSHOT</version>
             </dependency>
             <dependency>
